add luogu

+++ luogu/P2068.cpp
@@ -1,14 +1,14 @@
 #include <bits/stdc++.h>
 using namespace std;
 int n, m;
-int c[100020];
+long long c[100020];
 void change(int x, int y) {
 	for (int i = x; i <= n; i += i & -i) {
 		c[i] += y;
 	}
 }
-int query(int x) {
-	int re = 0;
+long long query(int x) {
+	long long re = 0;
 	for (int i = x; i > 0; i -= i & -i) {
 		re += c[i];
 	}
@@ -23,7 +23,7 @@ int main() {
 		if (o == 'x') {
 			change(x, y);
 		} else {
-			printf("%d\n", query(y) - query(x - 1));
+			printf("%lld\n", query(y) - query(x - 1));
 		}
 	}
 	return 0;
